refactor(junyeong/todoComment/#15): models 폴더 구조 변경
