+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "body-parser": "^1.20.1",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express.js": "^1.0.0",
